cleaned up branch and ready for submit
