--- Handige websites
@@ -1,3 +1,0 @@
-Custom dataset via pytorch:
-
-https://github.com/utkuozbulak/pytorch-custom-dataset-examples
