--- loadnetwork_filters.py
@@ -1,133 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Mon May 20 10:46:58 2019
-
-@author: RJ
-"""
-
-import torch
-import torchvision
-import torchvision.transforms as transforms
-import numpy as np
-import torch.nn as nn
-import torch.nn.functional as F
-import pandas as pd
-from PIL import Image
-
-#%% define parameters
-param_path = 'weights_filters.pth'     # path to model param
-test_path = 'SIGN/sign_mnist_test.csv'  # path to test csv
-
-N_classes = 26          # number of classes
-batch = 8               # batch size
-
-#%% define dataloader
-class SIGN(torch.utils.data.Dataset):
-    def __init__(self,csv_file,height,width, transforms=None):
-        self.data = pd.read_csv(csv_file)
-        self.labels = np.asarray(self.data.iloc[:, 0])
-        self.height = height
-        self.width = width
-        self.transforms = transforms
-
-    def __getitem__(self,index):
-        label = self.labels[index]
-        img_as_np = np.asarray(self.data.iloc[index][1:]).reshape(28,28).astype('uint8')
-        img_as_img = Image.fromarray(img_as_np)
-        img_as_img = img_as_img.convert('L')
-        if self.transforms is not None:
-            img_as_tensor = self.transforms(img_as_img)
-        return (img_as_tensor,label)
-    def __len__(self):
-       return len(self.data.index)
-   
-#%% load data
-transform = transforms.Compose([transforms.ToTensor()])
-testset = SIGN(test_path,28,28,transform)
-testloader = torch.utils.data.DataLoader(testset, batch_size=batch,
-                                         shuffle=False, num_workers=1)
-
-#%% define classes
-classes = ('A', 'B', 'C', 'D',
-           'E', 'F', 'G', 'H',
-           'I', 'J', 'K', 'L', 
-           'M', 'N', 'O', 'P',
-           'Q', 'R', 'S', 'T',
-           'U', 'V', 'W', 'X',
-           'Y', 'Z')
-
-#%% define neural network
-class Net(nn.Module):
-    def __init__(self):
-        super(Net, self).__init__()
-        # Conv2d(in_channels, out_channels, kernel_size, stride=1, padding=0, 
-        #        dilation=1, groups=1, bias=True, padding_mode='zeros')
-        self.conv1 = nn.Conv2d(1, 30, 3, padding=1)
-        self.conv2 = nn.Conv2d(30, 60, 3, padding=1)
-        # MaxPool2d(kernel_size, stride=None, padding=0, dilation=1, 
-        #           return_indices=False, ceil_mode=False)
-        self.pool = nn.MaxPool2d(2, 2)
-        self.conv3 = nn.Conv2d(60,90, 3, padding=1)
-        self.conv4 = nn.Conv2d(90,120, 3, padding=1)
-        self.pool = nn.MaxPool2d(2, 2)
-        # Linear(in_features, out_features, bias=True)
-        self.fc1 = nn.Linear(120 * 7 * 7, 240)
-        self.fc2 = nn.Linear(240, 80)
-        self.fc3 = nn.Linear(80, N_classes)
-
-    def forward(self, x):
-        x = F.relu(self.conv1(x))
-        x = self.pool(F.relu(self.conv2(x)))
-        x = F.relu(self.conv3(x))
-        x = self.pool(F.relu(self.conv4(x)))
-        x = x.view(-1, 120 * 7 * 7)
-        x = F.relu(self.fc1(x))
-        x = F.relu(self.fc2(x))
-        x = self.fc3(x)
-        return x
-    
-#%% testing network
-if __name__ == '__main__':
-    use_gpu = torch.cuda.is_available()
-    net = Net()
-    if use_gpu:
-        net = net.cuda()
-        device = torch.device("cuda")
-    net.load_state_dict(torch.load(param_path))
-#    net.load_state_dict(torch.load(param_path,map_location='cpu'))
-    net.eval()
-    
-    print('Start testing overall')
-    correct = 0
-    total = 0
-    with torch.no_grad():
-        for data in testloader:
-            images, labels = data
-            outputs = net(images.to(device))
-            _, predicted = torch.max(outputs.data, 1)
-            total += labels.size(0)
-            correct += (predicted == labels.to(device)).sum().item()
-
-    print('Accuracy of the network on the test images: %d %%' % (100 * correct / total))
-    
-    print('Start testing per class')
-    class_correct = list(0. for i in range(N_classes))
-    class_total = list(0. for i in range(N_classes))
-    with torch.no_grad():
-        for data in testloader:
-            images, labels = data
-            outputs = net(images.to(device))
-            _, predicted = torch.max(outputs, 1)
-            c = (predicted == labels.to(device)).squeeze()
-            for i in range(4):
-                label = labels[i]
-                class_correct[label] += c[i].item()
-                class_total[label] += 1
-
-
-    for i in range(N_classes):
-        if i == 9 or i == 25:   # skip J and Z since they are not included
-            pass
-        else:
-            print('Accuracy of %5s : %2d %%' % (classes[i], 100 * class_correct[i] / class_total[i]))
