--- loss_NLLLoss.txt
@@ -1,94 +1,0 @@
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-nan
-94.0219002366066
-88.637440264225
-87.77403950691223
-87.91298907995224
-88.09304916858673
-85.91355693340302
-86.1462008357048
-87.50679212808609
-87.2638652920723
-86.20816546678543
-84.65948742628098
-70.38477647304535
-42.561887845396996
-28.29255513101816
-20.017985079437494
-15.838128995150328
-12.954104475677013
-11.60364737920463
-8.530419858172536
-8.089939431287348
-7.051739163696766
-6.563062312081456
-5.5452524670399725
-4.91626523155719
