--- loss_mseloss.txt
@@ -1,4 +1,0 @@
-324.02670073509216
-322.97975301742554
-321.5477030277252
-319.78292965888977
