--- results/resultfs0lr0mo0extended.txt
@@ -1,25 +1,0 @@
-94.01840490797547
-100.0
-97.3913043478261
-100.0
-100.0
-99.56709956709956
-100.0
-78.4090909090909
-95.45454545454545
-100.0
-84.91620111731844
-100.0
-97.8021978021978
-74.46808510638297
-99.2063492063492
-100.0
-100.0
-100.0
-100.0
-75.75757575757575
-89.47368421052632
-95.34883720930233
-100.0
-98.57142857142857
-87.8048780487805
